ref(server setters and getters): Split up into individual files
